Reorganized how the routes are defined and added a test user to the setup script.
