+++ server/api/secure.js
@@ -1,32 +1,15 @@
 'use strict';
 
-const Hoek = require('hoek');
-
-
-exports.register = function (server, options, next) {
-  options = Hoek.applyToDefaults({ basePath: '' }, options);
-
-
-  server.route({
-    method: 'get',
-    path: options.basePath + '/secure',
-    config: {
-      auth: {
-        mode: 'required',
-        strategy: 'jwt'
-      }
-    },
-    handler: (request, reply) => {
-        reply({ message: 'You are secure.' });
-    }
-  });
-
-
-  next();
+exports.secure = {
+  description: 'Route to demonstrate how to call a secure route',
+  notes: 'This should be removed once everyone knows how to access secure routes.',
+  tags: ['api'],
+  auth: {
+    mode: 'required',
+    strategy: 'jwt'
+  },
+  handler: (request, reply) => {
+    reply({ message: 'You are secure.' });
+  }
 };
 
-
-exports.register.attributes = {
-  name: 'secure',
-  version: '1.0.0'
-};
