+++ server/config/manifest.js
@@ -44,11 +44,7 @@ const manifest = {
     } },
     { plugin: { register: './auth/auth', options: {}, } },
     { plugin: { register: './auth/token', options: {}, } },
-    { plugin: { register: './api/login', options: { basePath: '/api' }, } },
-    { plugin: { register: './api/register', options: { basePath: '/api' }, } },
-    { plugin: { register: './api/secure', options: { basePath: '/api' }, } },
-    { plugin: { register: './web/public', options: {}, } },
-    { plugin: { register: './web/home', options: {}, } },
+    { plugin: { register: './routes', options: { basePath: '/api' }, } },
   ]
 };
