--- server/web/home/index.js
@@ -1,20 +1,0 @@
-'use strict';
-
-exports.register = function (plugin, options, next) {
-
-  plugin.route({
-    method: 'GET',
-    path: '/',
-    handler: {
-      file: {
-        path: 'public/index.html'
-      }
-    }
-  });
-
-
-  next();
-};
-
-
-exports.register.attributes = { name: 'web/home' };
